Rework some things
