+++ .env.example
@@ -25,4 +25,6 @@ PUBLIC_DONATION_URL="https://ko-fi.com/Kaeso"
 PUBLIC_TOS_URL="https://github.com/EliteFarmers/Website/wiki/Terms-of-Service"
 PUBLIC_PRIVACY_URL="https://github.com/EliteFarmers/Website/wiki/Privacy-Policy"
 
+# These are used to show a popup telling users to apply for Elite Farmer
 PUBLIC_WEIGHT_REQ="10000"
+PUBLIC_ELITE_BADGE_ID="1"
