+++ src/components/stats/player/weight.svelte
@@ -1,12 +1,6 @@
 <script lang="ts">
-	import { onMount } from 'svelte';
 	import { getStatsContext } from '$lib/stores/stats.svelte';
 	import Profiles from './profiles.svelte';
-	import { PUBLIC_WEIGHT_REQ } from '$env/static/public';
-	import { page } from '$app/state';
-	import EliteToast from '$comp/stats/player/eliteToast.svelte';
-	import { toast } from 'svelte-sonner';
-	import { tick } from 'svelte';
 
 	const ctx = getStatsContext();
 
@@ -14,21 +8,6 @@
 	const rank = $derived(ctx.ranks?.misc?.farmingweight);
 	const rankText = $derived(rank !== -1 ? `#${rank}` : 'Unranked');
 	const weightStr = $derived(weightInfo?.totalWeight?.toLocaleString() ?? '0');
-
-	onMount(async () => {
-		await tick();
-
-		const eligible = (ctx.member.farmingWeight?.totalWeight ?? 0) >= Number(PUBLIC_WEIGHT_REQ);
-		const isOwnAccount = (page.data.session.ign ?? undefined) === ctx.ign;
-		const hasElite = ctx.account.badges?.some((badge) => badge.id !== undefined && badge.id === 1) ?? false;
-
-		if (isOwnAccount && eligible && !hasElite) {
-			// @ts-expect-error - Not updated for Svelte 5 yet
-			toast.custom(EliteToast, {
-				duration: Number.POSITIVE_INFINITY,
-			});
-		}
-	});
 </script>
 
 <div class="block">
